trade-opportunities-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
PyJWT==2.8.0
google-generativeai==0.7.2
duckduckgo-search==6.2.6
